경북대 BE_정수현 4주차 과제(2단계) (#290)
* feat: Category는 id를 따로 가짐

* chore: 사용하지 않는 메서드 제거

* fix: name column is duplicated

* feat: category name은 unique

* refactor: requestentitybody 부분 줄이기

* refactor: jackson objectmapper는 기본생성자만을 필요로 한다

* refactor: variable를 꺼냄

* fix: categoryid를 얻기

* docs: option feat

* feat: Option entity클래스

* feat: OptionRepository

* feat: OptionService getAllOption

* feat: OptionService createOption

* feat: Option update, delete

* feat: OptionController

* feat: OptionRequest valid

* feat: product updateInfo에 validate 추가

* fix: optionRepository Long 잘못 참조 해결

* chore: 중복 삭제

* chore: code convention

* test: optionTest

* feat: Product NotFound Error

* feat: member errorcode

* feat: OptionErrorCode 추가

* feat: exception들 추가

* feat: exception handler 추가

* feat: nameDuplicate Error

* feat: wishPathvariable로 삭제

* test: update test추가

* feat: product가 존재하면 category 지우지못함

* feat: option이 1개 이하면 지울 수 없음

* chore: 오타 수정

* refactor: dto는 record로 작성

* refactor: dto는 record 클래스

* chore: code convention

* refactor: productrequest를 create와 update로 쪼갬

* feat: product 만들 때에는 option을 무조건 하나 만들어야함.

+++ src/main/java/gift/category/CategoryService.java
@@ -1,8 +1,10 @@
 package gift.category;
 
 import gift.category.model.Category;
-import gift.category.model.CategoryRequestDto;
-import gift.category.model.CategoryResponseDto;
+import gift.category.model.CategoryRequest;
+import gift.category.model.CategoryResponse;
+import gift.common.exception.CategoryException;
+import gift.product.ProductRepository;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -11,41 +13,47 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class CategoryService {
 
+    private final ProductRepository productRepository;
     private final CategoryRepository categoryRepository;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(ProductRepository productRepository, CategoryRepository categoryRepository) {
+        this.productRepository = productRepository;
         this.categoryRepository = categoryRepository;
     }
 
     @Transactional(readOnly = true)
-    public List<CategoryResponseDto> getAllCategories(Pageable pageable) {
+    public List<CategoryResponse> getAllCategories(Pageable pageable) {
         return categoryRepository.findAll(pageable)
-            .map(CategoryResponseDto::from)
+            .map(CategoryResponse::from)
             .getContent();
     }
 
     @Transactional(readOnly = true)
-    public CategoryResponseDto getCategoryById(Long id) {
+    public CategoryResponse getCategoryById(Long id) throws CategoryException {
         Category category = categoryRepository.findById(id)
-            .orElseThrow(() -> new IllegalArgumentException("Category 가 잘못되었습니다."));
-        return CategoryResponseDto.from(category);
+            .orElseThrow(() -> new CategoryException(CategoryErrorCode.NOT_FOUND));
+        return CategoryResponse.from(category);
     }
 
     @Transactional
-    public Long insertCategory(CategoryRequestDto categoryRequestDto) {
-        Category category = categoryRepository.save(categoryRequestDto.toEntity());
+    public Long insertCategory(CategoryRequest categoryRequest) {
+        Category category = new Category(categoryRequest.name(), categoryRequest.color(), categoryRequest.imageUrl(), categoryRequest.description());
+        category = categoryRepository.save(category);
         return category.getId();
     }
 
     @Transactional
-    public void updateCategory(CategoryRequestDto categoryRequestDto, Long id) {
+    public void updateCategory(CategoryRequest categoryRequest, Long id) throws CategoryException {
         Category category = categoryRepository.findById(id)
-            .orElseThrow(() -> new IllegalArgumentException("Category 가 잘못되었습니다."));
-        category.updateInfo(categoryRequestDto.toEntity());
+            .orElseThrow(() -> new CategoryException(CategoryErrorCode.NOT_FOUND));
+        category.updateInfo(categoryRequest.name(), categoryRequest.color(), categoryRequest.imageUrl(), categoryRequest.description());
     }
 
     @Transactional
-    public void deleteCategory(Long id) {
+    public void deleteCategory(Long id) throws CategoryException {
+        if(productRepository.existsByCategoryId(id)){
+            throw new CategoryException(CategoryErrorCode.CAN_NOT_DELETE);
+        }
         categoryRepository.deleteById(id);
     }
 }
